refactor: credit의 타입을 Long으로 통일

+++ src/main/java/doritos/doriroom/diary/dto/response/DiaryResponseDto.java
@@ -29,7 +29,7 @@ public record DiaryResponseDto(
     int likes,
 
     @Schema(description = "지급된 크레딧", example = "3")
-    int credit,
+    Long credit,
 
     @Schema(description = "방문 날짜", example = "2025-08-10")
     String visitedAt,
@@ -43,7 +43,7 @@ public record DiaryResponseDto(
     @Schema(description = "축제 정보")
     EventInfoResponseDto eventInfo
 ) {
-    public static DiaryResponseDto from(Diary diary, int totalCredit, User user, Event event) {
+    public static DiaryResponseDto from(Diary diary, Long totalCredit, User user, Event event) {
         return new DiaryResponseDto(
             diary.getDiaryId(),
             diary.getContent(),
